refactor: extract MiddleTruncateName, adopt in file browser rows and cards
Middle truncation moves from a dashboard-page private component to components/dashboard/MiddleTruncateName.tsx and replaces CSS end-truncation in FileRow/FileCard, so burst siblings and extensions stay distinguishable in the file browser too. Convention documented in docs/ai/conventions.md. E2E filename locators gain .first() (the component renders an sr-only full copy plus the aria-hidden fitted copy).
